Add Razor Login page
