+++ Solution/Localization.Web.AspNetCore.Sample/Views/Home/Contact.cshtml
@@ -16,23 +16,23 @@
                     <hr />
                     @Html.ValidationSummary(true, "", new { @class = "text-danger" })
                     <div class="form-group">
-                        @Html.LabelFor(m => m.UserName, new { @class = "col-md-3 control-label" })
+                        @Html.LabelFor(m => m.Input.UserName, new { @class = "col-md-3 control-label" })
                         <div class="col-md-8">
-                            @Html.TextBoxFor(m => m.UserName, new { @class = "form-control" })
+                            @Html.TextBoxFor(m => m.Input.UserName, new { @class = "form-control" })
                         </div>
                     </div>
                     <div class="form-group">
-                        @Html.LabelFor(m => m.Password, new { @class = "col-md-3 control-label" })
+                        @Html.LabelFor(m => m.Input.Password, new { @class = "col-md-3 control-label" })
                         <div class="col-md-8">
-                            @Html.PasswordFor(m => m.Password, new { @class = "form-control" })
-                            @Html.ValidationMessageFor(m => m.Password, "", new { @class = "text-danger" })
+                            @Html.PasswordFor(m => m.Input.Password, new { @class = "form-control" })
+                            @Html.ValidationMessageFor(m => m.Input.Password, "", new { @class = "text-danger" })
                         </div>
                     </div>
                     <div class="form-group">
                         <div class="col-md-offset-3 col-md-9">
                             <div class="checkbox">
-                                @Html.CheckBoxFor(m => m.RememberMe)
-                                @Html.LabelFor(m => m.RememberMe)
+                                @Html.CheckBoxFor(m => m.Input.RememberMe)
+                                @Html.LabelFor(m => m.Input.RememberMe)
                             </div>
                         </div>
                     </div>
