--- Solution/Localization.Web.AspNetCore.Sample/Views/Home/Login.cshtml
@@ -1,8 +1,0 @@
-﻿@model object
-
-@{
-    ViewBag.Title = "title";
-    Layout = "_Layout";
-}
-
-<h2>LOGIN</h2>
